Changes for the IP transfer functionality requested by Ryan Baltrush
If the user doesn't have a hard copy of the document, he will still be
able to affirm the AD online (this will prevent the process from
halting) but he won't be paid until TC receives the copy.
If the user have a hard copy of the document associated, he will be able
to affirm the AD online and he won't need to send again a copy.